// src/types/app.ts
export interface ThemeColor extends OtherColor {
  primary: string
}
// ...
export type DropdownKey = 'closeCurrent' | 'closeOther' | 'closeLeft' | 'closeRight' | 'closeAll';

/**
 * I18n namespace
 *
 * Locales export type
 */

export type LangType = 'en-US' | 'zh-CN';

export interface LangOption {
  label: string
  key: LangType
  [k: string]: unknown
}

export type I18nRouteKey = Exclude<RouteKey, 'root' | 'not-found'>;

export interface FormMsg {
  required: string
  invalid: string
}

export interface I18nSchema {
  system: {
    title: string
    updateTitle: string
    updateContent: string
    updateConfirm: string
    updateCancel: string
  }
  common: {
    action: string
    add: string
    addSuccess: string
    backToHome: string
    batchDelete: string
    cancel: string
    close: string
    check: string
    expandColumn: string
    columnSetting: string
    config: string
    confirm: string
    delete: string
    deleteSuccess: string
    confirmDelete: string
    edit: string
    warning: string
    error: string
    index: string
    keywordSearch: string
    logout: string
    logoutConfirm: string
    lookForward: string
    modify: string
    modifySuccess: string
    noData: string
    operate: string
    pleaseCheckValue: string
    refresh: string
    reset: string
    search: string
    switch: string
    tip: string
    trigger: string
    update: string
    updateSuccess: string
    userCenter: string
    yesOrNo: {
      yes: string
      no: string
    }
  }
  request: {
    logout: string
    logoutMsg: string
    logoutWithModal: string
    logoutWithModalMsg: string
    refreshToken: string
    tokenExpired: string
  }
  theme: {
    themeSchema: { title: string } & Record<ThemeScheme, string>
    grayscale: string
    colorWeakness: string
    layoutMode: { title: string, reverseHorizontalMix: string } & Record<ThemeLayoutMode, string>
    recommendColor: string
    recommendColorDesc: string
    themeColor: {
      title: string
      followPrimary: string
    } & ThemeColor
    scrollMode: { title: string } & Record<ThemeScrollMode, string>
    page: {
      animate: string
      mode: { title: string } & Record<ThemePageAnimateMode, string>
    }
    fixedHeaderAndTab: string
    header: {
      height: string
      breadcrumb: {
        visible: string
        showIcon: string
      }
      multilingual: {
        visible: string
      }
    }
    tab: {
      visible: string
      cache: string
      height: string
      mode: { title: string } & Record<ThemeTabMode, string>
    }
    sidebar: {
      inverted: string
      width: string
      collapsedWidth: string
      mixWidth: string
      mixCollapsedWidth: string
      mixChildMenuWidth: string
    }
    footer: {
      visible: string
      fixed: string
      height: string
      right: string
    }
    watermark: {
      visible: string
      text: string
    }
    themeDrawerTitle: string
    pageFunTitle: string
    resetCacheStrategy: { title: string } & Record<ResetCacheStrategy, string>
    configOperation: {
      copyConfig: string
      copySuccessMsg: string
      resetConfig: string
      resetSuccessMsg: string
    }
  }
  route: Record<I18nRouteKey, string>
  page: {
    login: {
      common: {
        loginOrRegister: string
        userNamePlaceholder: string
        phonePlaceholder: string
        codePlaceholder: string
        passwordPlaceholder: string
        confirmPasswordPlaceholder: string
        codeLogin: string
        confirm: string
        back: string
        validateSuccess: string
        loginSuccess: string
        welcomeBack: string
      }
      pwdLogin: {
        title: string
        rememberMe: string
        forgetPassword: string
        register: string
        otherAccountLogin: string
        otherLoginMode: string
        superAdmin: string
        admin: string
        user: string
      }
      codeLogin: {
        title: string
        getCode: string
        reGetCode: string
        sendCodeSuccess: string
        imageCodePlaceholder: string
      }
      register: {
        title: string
        agreement: string
        protocol: string
        policy: string
      }
      resetPwd: {
        title: string
      }
      bindWeChat: {
        title: string
      }
    }
    home: {
      branchDesc: string
      greeting: string
      weatherDesc: string
      projectCount: string
      todo: string
      message: string
      downloadCount: string
      registerCount: string
      schedule: string
      study: string
      work: string
      rest: string
      entertainment: string
      visitCount: string
      turnover: string
      dealCount: string
      projectNews: {
        title: string
        moreNews: string
        desc1: string
        desc2: string
        desc3: string
        desc4: string
        desc5: string
      }
      creativity: string
    }
  }
  form: {
    required: string
    userName: FormMsg
    phone: FormMsg
    pwd: FormMsg
    confirmPwd: FormMsg
    code: FormMsg
    email: FormMsg
  }
  dropdown: Record<DropdownKey, string>
  icon: {
    themeConfig: string
    themeSchema: string
    lang: string
    fullscreen: string
    fullscreenExit: string
    reload: string
    collapse: string
    expand: string
    pin: string
    unpin: string
  }
  dataTable: {
    itemCount: string
  }
}

export type GetI18nKey<T extends Record<string, any>, K extends keyof T = keyof T> = K extends string
  ? T[K] extends Record<string, unknown>
    ? `${K}.${GetI18nKey<T[K]>}`
    : K
  : never;

export type I18nKey = GetI18nKey<I18nSchema>;